create a dateGetter util class
move GetStartingDate from CurrencyBeaconApiClient into dataGetter class since curService class is also need to invoke the function
